requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
whitenoise==6.11.0
gunicorn==23.0.0
Pillow==12.0.0

# Additional packages for AI and voice processing
openai==1.54.0
requests==2.31.0
python-dotenv==1.0.0

# For voice processing (optional)
# SpeechRecognition==3.10.0
# gTTS==2.5.0
# pydub==0.25.1

# For document processing (optional)
# PyMuPDF==1.23.0
# python-docx==1.1.0
# pandas==2.1.0
# xlrd==2.0.1
